requirements.txt
gitpython==3.1.12
requests
redis==3.5.3
boto3==1.16.57
jinja2==2.11.2
numpy
sklearn